requirements.txt
streamlit>=1.41,<2.0
pydantic>=2.7,<3.0
email-validator>=2.2,<3.0
python-dotenv>=1.0,<2.0
openai>=2.24.0,<3.0
